+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - #11 support additional format options as per json schema (uri, hostname, ipv4, ipv6)
+- Fix the `maxItems` bug and add additional test cases - #16
 
 # v0.0.3